chore: move function to another file
