+++ src/utils.ts
@@ -10,13 +10,3 @@ export function time2String(time: number) {
     timeArr.push(second.toString().padStart(2, "0"));
     return timeArr.join(":");
 }
-
-let DebounceTimer: any = null;
-
-export function debounce<T extends Function>(cb: T, wait = 20) {
-    let callable = (...args: any) => {
-        clearTimeout(DebounceTimer);
-        DebounceTimer = setTimeout(() => cb(...args), wait);
-    };
-    return <T>(<any>callable);
-}
